requirements.txt
appdirs>=1.4.4
numpy>=1.26.4,<2.0.0
h5py>3.7.0
matplotlib>=3.5.2
pandas>=1.4.3
tqdm>=4.66.3
scipy
pyyaml>=6.0
pint
xarray
netCDF4>=1.7.0
h5rdmtoolbox>=1.6.0